editoast: rename views::projects to view::project for consistency
