+++ editoast/src/views/scenario/macro_nodes.rs
@@ -30,8 +30,8 @@ use crate::views::AuthorizationError;
 use crate::views::pagination::PaginatedList;
 use crate::views::pagination::PaginationQueryParams;
 use crate::views::pagination::PaginationStats;
-use crate::views::projects::ProjectError;
-use crate::views::projects::ProjectIdParam;
+use crate::views::project::ProjectError;
+use crate::views::project::ProjectIdParam;
 use crate::views::scenario::ScenarioError;
 use crate::views::scenario::ScenarioIdParam;
 use crate::views::study::StudyError;
